pyproject.toml
[project]
name = "slack-ai-agent-v2"
version = "0.2.0"
description = "Clean Slack AI agent with RAG and CSV analysis capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "slack-bolt>=1.23.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.2.8",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langgraph>=0.2.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']